requirements.txt
numpy==1.19.5
matplotlib==3.4.2
scipy==1.6.2
streamlit==0.84.2
tensorflow-cpu==2.5.0
keras_nightly==2.5.0.dev2021032900
keras==2.4.3
Pillow==8.3.1
